pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["Vitaliy Pavlov <v-2841@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <=3.12"
django = "^5.0.6"
ip2location = "^8.10.2"
djlint = "^1.34.1"
django-debug-toolbar = "^4.4.2"
pillow = "^10.3.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
